Recreate patch for updated library
Due to some changes in the library, the old patch file could no longer
be applied. In general, the changes induced by the patch are sill mostly
the same, but the line numbers have changed. In addition, some new files
were added by the update, whose header was patched as well.